requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
autoflake==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.2.0
black==23.11.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
executing==2.0.1
fastapi==0.104.0
fastapi-cache2==0.2.1
fastapi-versioning==0.10.0
flake8==6.1.0
flower==2.0.1
frozenlist==1.4.0
greenlet==3.0.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
humanize==4.8.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.17.2
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.5.0
jupyter_core==5.5.0
kombu==5.3.3
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nodeenv==1.8.0
orjson==3.9.10
outcome==1.3.0.post0
packaging==23.2
parso==0.8.3
passlib==1.7.4
pathspec==0.11.2
pendulum==2.1.2
Pillow==10.1.0
platformdirs==3.11.0
playwright==1.39.0
pluggy==1.3.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyee==11.0.1
pyflakes==3.1.0
Pygments==2.16.1
pyright==1.1.335
PySocks==1.7.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3.post1
pytzdata==2020.1
pywin32==306; platform_system=="Windows"
PyYAML==6.0.1
pyzmq==25.1.1
redis==4.6.0
requests==2.31.0
rsa==4.9
selenium==4.15.2
sentry-sdk==1.34.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqladmin==0.15.2
SQLAlchemy==2.0.22
stack-data==0.6.3
starlette==0.27.0
tornado==6.3.3
traitlets==5.13.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
undetected-chromedriver==3.5.4
urllib3==1.26.18
uvicorn==0.23.2
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.9
websockets==12.0
wsproto==1.2.0
WTForms==3.0.1
yarl==1.9.2
openai==0.27.8
